requirements.txt
flask
python-dotenv
groq
transformers
pillow
huggingface_hub
torch
